+++ src_ts/components/pages/interventions/interventions-module.ts
@@ -8,8 +8,8 @@ import '@polymer/paper-menu-button/paper-menu-button.js';
 import CONSTANTS from '../../../config/app-constants';
 import EndpointsMixin from '../../endpoints/endpoints-mixin';
 import ScrollControlMixin from '../../common/mixins/scroll-control-mixin';
-import ModuleMainElCommonFunctionalityMixin from '../mixins/module-common-mixin';
-import ModuleRoutingMixin from '../mixins/module-routing-mixin';
+import ModuleMainElCommonFunctionalityMixin from '../../common/mixins/module-common-mixin';
+import ModuleRoutingMixin from '../../common/mixins/module-routing-mixin';
 import '../../common/components/page-content-header.js';
 import '../../styles/page-content-header-slotted-styles.js';
 import '../../common/components/etools-error-messages-box.js';
