+++ src_ts/components/pages/reports/reports-module.ts
@@ -18,8 +18,8 @@ import './components/report-status';
 import './components/report-rating-dialog';
 import './components/report-reject-dialog';
 import {GenericObject, User} from '@unicef-polymer/etools-types';
-import ModuleMainElCommonFunctionalityMixin from '../mixins/module-common-mixin';
-import ModuleRoutingMixin from '../mixins/module-routing-mixin';
+import ModuleMainElCommonFunctionalityMixin from '../../common/mixins/module-common-mixin';
+import ModuleRoutingMixin from '../../common/mixins/module-routing-mixin';
 import ScrollControlMixin from '../../common/mixins/scroll-control-mixin';
 import {pageLayoutStyles} from '../../styles/page-layout-styles';
 import {SharedStyles} from '../../styles/shared-styles';
